Obsoleted ID Classes, fixed various errors
Moved away from the ID Class system. Everything has its own id field as an Integer now, and the UniqueConstraints annotation is used to ensure uniqueness.

Updated the Grade entity to have its id as an Integer. The difference really doesn't matter for this project, but since everything else was an Integer, it's easier keep it consistent.

Fixed various issues with the GradebookServiceImpl file as well.